requirements.txt
cycler==0.11.0
dataclasses==0.6
einops==0.4.1
fonttools==4.31.2
future==0.18.2
imageio==2.9.0
kiwisolver==1.4.2
matplotlib==3.5.1
networkx==2.6.3
opencv-python==4.5.5.64
packaging==21.3
pillow==9.1.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytorch-msssim==0.2.1
pywavelets==1.3.0
pyyaml==5.4.1
scikit-image==0.15.0
scipy==1.7.3
tifffile==2021.11.2
timm==0.5.4
torch==1.7.0
torchvision==0.8.1
tqdm==4.36.1
typing-extensions==4.1.1
